ping settings
AooClient and AooServer now exchange pings/pongs to detect broken connections.
The ping settings are similar to TCP keepalive and are managed with the AooPingSettings struct.

The same mechanism is also used to detect broken peer UDP connections; this is reported with the new AooEventPeerState event.

The initial query phase (interval + timeout) for the UDP server and peers is still managed with AOO_CLIENT_QUERY_INTERVAL and AOO_CLIENT_QUERY_TIMEOUT;
the UDP server pings are still managed with AOO_CLIENT_PING_INTERVAL.