aoo: add "channel" member to AooStreamMessage
Users can now associate stream messages with particular channels.
This is also useful for pure message streams, as it allows users to group messages into channels.
For example, they can be used as MIDI tracks (as opposed to channels).

It also gives meaning to the "numChannels" member in AooCodecNull, as it tells the user the total number of message channels.
Note that the channel count is not really enforced (yet), i.e. you can send stream messages with "out-of-range" channel numbers.

NB: we now limit stream message data to 65535 bytes!